Refactor reusable UI components
Extracted custom table cells and text field to dedicated files, improving modularity and code organization

- Moved CategoryCell, ScheduleCell, and SpacedTextField from controllers to Views/ as standalone classes
- Simplified TrackerCategorySelectionViewController and TrackerScheduleSelectionViewController by referencing new view components
- Removed duplicated cell and text field code from respective controllers
